+++ CHANGES.rst
@@ -6,6 +6,8 @@
 - Keep notebook metadata on push to prevent deleting kernel information
 - Added sctoolbox as default kernel to RNA & ATAC notebooks
 - Added check of column validity to tools.marker_genes.run_DESeq2() (#134)
+- Increase test coverage for plotting functions (#126)
+- Apply fixes to bugs found by increasing the test coverage.
 
 Changes to ATAC notebooks (0.1.1.2 - 05-06-2023):
 - Fixed assembling atac notebook 01